+++ app/src/main/kotlin/ac/mdiq/podcini/ui/screens/FeedDetailsScreen.kt
@@ -587,7 +587,7 @@ fun FeedDetailsScreen() {
     fun MyTopAppBar(displayUpArrow: Boolean) {
         val context = LocalContext.current
         var expanded by remember { mutableStateOf(false) }
-        TopAppBar(title = { Text("") }, modifier = Modifier.height(40.dp),
+        TopAppBar(title = { Text("") }, 
             navigationIcon = if (displayUpArrow) {
                 { IconButton(onClick = { if (mainNavController.previousBackStackEntry != null) mainNavController.popBackStack()
                 }) { Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back") } }