+++ app/src/main/kotlin/ac/mdiq/podcini/ui/screens/OnlineSearchScreen.kt
@@ -279,7 +279,7 @@ fun OnlineSearchScreen() {
     @OptIn(ExperimentalMaterial3Api::class)
     @Composable
      fun MyTopAppBar() {
-        TopAppBar(modifier = Modifier.height(40.dp), title = { SearchBarRow(R.string.search_podcast_hint) { queryText ->
+        TopAppBar( title = { SearchBarRow(R.string.search_podcast_hint) { queryText ->
             if (queryText.isBlank()) return@SearchBarRow
             if (queryText.matches("http[s]?://.*".toRegex())) vm.addUrl(queryText)
             else {