+++ app/src/main/kotlin/ac/mdiq/podcini/ui/screens/StatisticsScreen.kt
@@ -104,7 +104,7 @@ fun StatisticsScreen() {
     @Composable
     fun MyTopAppBar() {
         var expanded by remember { mutableStateOf(false) }
-        TopAppBar(title = { Text(stringResource(R.string.statistics_label)) }, modifier = Modifier.height(40.dp),
+        TopAppBar(title = { Text(stringResource(R.string.statistics_label)) }, 
             navigationIcon = { IconButton(onClick = {
 //            if (mainNavController.previousBackStackEntry != null) mainNavController.popBackStack()
                 MainActivity.openDrawer()